Refactor editor components
